+++ nautilus_trader/backtest/engine.pyx
@@ -66,7 +66,7 @@ from nautilus_trader.model.objects cimport Currency
 from nautilus_trader.model.orderbook.data cimport OrderBookData
 from nautilus_trader.portfolio.base cimport PortfolioFacade
 from nautilus_trader.system.kernel cimport NautilusKernel
-from nautilus_trader.trading.strategy cimport TradingStrategy
+from nautilus_trader.trading.strategy cimport Strategy
 from nautilus_trader.trading.trader cimport Trader
 
 
@@ -606,11 +606,11 @@ cdef class BacktestEngine:
         # Checked inside trader
         self.kernel.trader.add_actors(actors)
 
-    def add_strategy(self, strategy: TradingStrategy) -> None:
+    def add_strategy(self, strategy: Strategy) -> None:
         # Checked inside trader
         self.kernel.trader.add_strategy(strategy)
 
-    def add_strategies(self, strategies: List[TradingStrategy]) -> None:
+    def add_strategies(self, strategies: List[Strategy]) -> None:
         # Checked inside trader
         self.kernel.trader.add_strategies(strategies)
 
@@ -916,7 +916,7 @@ cdef class BacktestEngine:
         cdef list time_events = []  # type: list[TimeEventHandler]
         cdef:
             Actor actor
-            TradingStrategy strategy
+            Strategy strategy
             cdef TimeEventHandler event_handler
         for actor in self.kernel.trader.actors_c():
             time_events += actor.clock.advance_time(now_ns)