+++ nautilus_trader/examples/strategies/blank.py
@@ -15,7 +15,7 @@
 
 from typing import Dict
 
-from nautilus_trader.config import TradingStrategyConfig
+from nautilus_trader.config import StrategyConfig
 from nautilus_trader.core.data import Data
 from nautilus_trader.core.message import Event
 from nautilus_trader.model.data.bar import Bar
@@ -23,10 +23,10 @@ from nautilus_trader.model.data.tick import QuoteTick
 from nautilus_trader.model.data.tick import TradeTick
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.instruments.base import Instrument
-from nautilus_trader.trading.strategy import TradingStrategy
+from nautilus_trader.trading.strategy import Strategy
 
 
-class MyStrategyConfig(TradingStrategyConfig):
+class MyStrategyConfig(StrategyConfig):
     """
     Configuration for ``MyStrategy`` instances.
 
@@ -43,7 +43,7 @@ class MyStrategyConfig(TradingStrategyConfig):
     instrument_id: str
 
 
-class MyStrategy(TradingStrategy):
+class MyStrategy(Strategy):
     """
     A blank template strategy.