+++ nautilus_trader/examples/strategies/ema_cross_bracket.py
@@ -17,7 +17,7 @@ from decimal import Decimal
 from typing import Dict, Optional
 
 from nautilus_trader.common.logging import LogColor
-from nautilus_trader.config import TradingStrategyConfig
+from nautilus_trader.config import StrategyConfig
 from nautilus_trader.core.data import Data
 from nautilus_trader.core.message import Event
 from nautilus_trader.indicators.atr import AverageTrueRange
@@ -28,14 +28,14 @@ from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.instruments.base import Instrument
 from nautilus_trader.model.orders.list import OrderList
-from nautilus_trader.trading.strategy import TradingStrategy
+from nautilus_trader.trading.strategy import Strategy
 
 
 # *** THIS IS A TEST STRATEGY WITH NO ALPHA ADVANTAGE WHATSOEVER. ***
 # *** IT IS NOT INTENDED TO BE USED TO TRADE LIVE WITH REAL MONEY. ***
 
 
-class EMACrossBracketConfig(TradingStrategyConfig):
+class EMACrossBracketConfig(StrategyConfig):
     """
     Configuration for ``EMACrossBracket`` instances.
 
@@ -70,7 +70,7 @@ class EMACrossBracketConfig(TradingStrategyConfig):
     trade_size: Decimal
 
 
-class EMACrossBracket(TradingStrategy):
+class EMACrossBracket(Strategy):
     """
     A simple moving average cross example strategy.