+++ tests/unit_tests/live/test_live_risk_engine.py
@@ -37,7 +37,7 @@ from nautilus_trader.model.identifiers import Venue
 from nautilus_trader.model.objects import Quantity
 from nautilus_trader.msgbus.bus import MessageBus
 from nautilus_trader.portfolio.portfolio import Portfolio
-from nautilus_trader.trading.strategy import TradingStrategy
+from nautilus_trader.trading.strategy import Strategy
 from tests.test_kit.mocks.exec_clients import MockExecutionClient
 from tests.test_kit.stubs.component import TestComponentStubs
 from tests.test_kit.stubs.events import TestEventStubs
@@ -159,7 +159,7 @@ class TestLiveRiskEngine:
             config=LiveRiskEngineConfig(qsize=1),
         )
 
-        strategy = TradingStrategy()
+        strategy = Strategy()
         strategy.register(
             trader_id=self.trader_id,
             portfolio=self.portfolio,
@@ -207,7 +207,7 @@ class TestLiveRiskEngine:
             config=LiveRiskEngineConfig(qsize=1),
         )
 
-        strategy = TradingStrategy()
+        strategy = Strategy()
         strategy.register(
             trader_id=self.trader_id,
             portfolio=self.portfolio,
@@ -278,7 +278,7 @@ class TestLiveRiskEngine:
         # Arrange
         self.risk_engine.start()
 
-        strategy = TradingStrategy()
+        strategy = Strategy()
         strategy.register(
             trader_id=self.trader_id,
             portfolio=self.portfolio,
@@ -320,7 +320,7 @@ class TestLiveRiskEngine:
         # Arrange
         self.risk_engine.start()
 
-        strategy = TradingStrategy()
+        strategy = Strategy()
         strategy.register(
             trader_id=self.trader_id,
             portfolio=self.portfolio,